// apps/host/src/types/hyperliquid.ts
export interface Leverage {
    type: "cross" | "isolated";
    value: number;
    rawUsd?: string;
}